// src/enums/characterClass.enum.ts
// Enum pour les classes de personnages
export enum CharacterClass {
    Warrior = 'Warrior',
    Deathknight = 'Deathknight',
    Paladin = 'Paladin',
    Monk = 'Monk',
    Priest = 'Priest',
    Mage = 'Mage',
    Druid = 'Druid',
    Rogue = 'Rogue',
    Hunter = 'Hunter',
    DemonHunter = 'Demon Hunter',
    Warlock = 'Warlock',
    Evoker = 'Evoker',
    Shaman = 'Shaman',
}